Moved GlobalMatrixAssemblerSkyline to Solvers because 1) it belongs there, 2) it can now work with embedding classes form FEM project, instead of Discretization.
