+++ ISAAR.MSolve.FEM/Elements/cohesive_shell_to_hexaCopyGetEmbeRAM_11_tlk.cs
@@ -18,6 +18,7 @@ using System.Linq;
 using System.Text;
 using ISAAR.MSolve.Discretization;
 using ISAAR.MSolve.Discretization.Interfaces;
+using IEmbeddedElement = ISAAR.MSolve.FEM.Interfaces.IEmbeddedElement;
 
 namespace ISAAR.MSolve.FEM.Elements
 {
